move Navigation component to @/components/app subfolder
